improved collision detection

+++ map/PowerArmor.java
@@ -3,7 +3,7 @@ import java.awt.Graphics;
 public class PowerArmor extends PowerUp{
 	
 
-	public PowerArmor(Handler handler,float x,float y,int width, int height){
+	public PowerArmor(Handler handler,int x,int y,int width, int height){
 		super(handler,x,y,width,height);
 	}
 
@@ -12,6 +12,6 @@ public class PowerArmor extends PowerUp{
 	}
 
 	public void render(Graphics g){
-		g.drawImage(Assets.powerArmor ,(int)x ,(int)y ,width ,height ,null);
+		g.drawImage(Assets.powerArmor ,x ,y ,width ,height ,null);
 	}
 }
