+++ map/TimeSkip.java
@@ -3,7 +3,7 @@ import java.awt.Graphics;
 public class TimeSkip extends PowerUp{
 	
 
-	public TimeSkip(Handler handler,float x,float y,int width, int height){
+	public TimeSkip(Handler handler,int x,int y,int width, int height){
 		super(handler,x,y,width,height);
 	}
 
@@ -12,6 +12,6 @@ public class TimeSkip extends PowerUp{
 	}
 
 	public void render(Graphics g){
-		g.drawImage(Assets.timeSkip ,(int)x ,(int)y ,width ,height ,null);
+		g.drawImage(Assets.timeSkip ,x ,y ,width ,height ,null);
 	}
 }
