+++ map/WallDestroyer.java
@@ -3,7 +3,7 @@ import java.awt.Graphics;
 public class WallDestroyer extends PowerUp{
 	
 
-	public WallDestroyer(Handler handler,float x,float y,int width, int height){
+	public WallDestroyer(Handler handler,int x,int y,int width, int height){
 		super(handler,x,y,width,height);
 	}
 
@@ -12,6 +12,6 @@ public class WallDestroyer extends PowerUp{
 	}
 
 	public void render(Graphics g){
-		g.drawImage(Assets.wallDestroyer ,(int)x ,(int)y ,width ,height ,null);
+		g.drawImage(Assets.wallDestroyer ,x ,y ,width ,height ,null);
 	}
 }
